+++ src/stores/content.ts
@@ -1,5 +1,5 @@
 import { defineStore } from 'pinia'
-import { DEFAULT_TEXT, CURRENT_CONTENT, HAVE_DATE, HAVE_WATERMARK, CURRENT_THEME, CURRENT_SIZE, TEXT_ALIGN, WRAPPER_MARGIN, FONT_FAMILY, FONT_FAMILY_ARR, THEME_ARR, SIZES_ARR } from './../helper/constant'
+import { DEFAULT_TEXT, CURRENT_CONTENT, HAVE_DATE, HAVE_WATERMARK, CURRENT_THEME, CURRENT_SIZE, TEXT_ALIGN, WRAPPER_MARGIN, FONT_FAMILY, FONT_FAMILY_ARR } from './../helper/constant'
 
 const VALID_FONT_IDS = new Set(FONT_FAMILY_ARR.map((item) => item.id))
 
@@ -17,7 +17,7 @@ export const useContentStore = defineStore({
       isWithDate: !!localStorage.getItem(HAVE_DATE),
       isWithWatermark: !!localStorage.getItem(HAVE_WATERMARK),
       content: localStorage.getItem(CURRENT_CONTENT) || DEFAULT_TEXT,
-      currentTheme: localStorage.getItem(CURRENT_THEME) || THEME_ARR[0].id,
+      currentTheme: localStorage.getItem(CURRENT_THEME) || 'notion',
       currentSize: localStorage.getItem(CURRENT_SIZE) || defaultSizeIdx,
       textAlign: localStorage.getItem(TEXT_ALIGN) || 'left',
       wrapperMargin: localStorage.getItem(WRAPPER_MARGIN) || 'standard',